refactor: change module & variables naming
